+++ docs/CHANGELOG.rst
@@ -6,6 +6,7 @@ Version 0.14.0dev (unreleased)
 
 * Upgrade default pip version to 1.3.1
 * Add support for Vagrant 1.1 providers in functional tests
+* Add ``fabtools.system.cpus`` to get the host's CPU count
 
 Version 0.13.0 (2013-03-15)
 ---------------------------